Starting using loaders from `loaders.gl`. Remove local loaders.
