+++ modules/builder/package.json
@@ -14,6 +14,7 @@
     "src"
   ],
   "dependencies": {
+    "@loaders.gl/gltf": "^0.4.8",
     "@turf/turf": "^5.1.6",
     "base64-js": "^1.3.0",
     "math.gl": "^2.0.0",
