+++ modules/parser/package.json
@@ -17,6 +17,7 @@
     "text-encoding": false
   },
   "dependencies": {
+    "@loaders.gl/gltf": "^0.4.8",
     "base64-js": "^1.3.0",
     "color": "^3.0.0",
     "math.gl": "^2.2.0",
