+++ modules/parser/src/index.js
@@ -40,10 +40,5 @@ export {parseStreamVideoMessage} from './parsers/parse-stream-video-message';
 export {default as lidarPointCloudWorker} from './workers/lidar-point-cloud-worker';
 export {default as streamDataWorker} from './workers/stream-data-worker';
 
-// LOADERS
-// Public methods
+// Loaders
 export {parseBinaryXVIZ} from './loaders/xviz-loader/xviz-binary-loader';
-export {parseGLB, _GLBDecoder, _unpackGLBBuffers, _unpackJsonArrays} from './loaders/glb-loader';
-
-export {default as PLYParser} from './loaders/ply-loader/ply-parser';
-export {loadBinary, parsePLY, generateNormals, normalizeXYZ} from './loaders/ply-loader/ply-loader';
