+++ modules/parser/webpack/webpack.config.js
@@ -34,5 +34,10 @@ module.exports = {
         ]
       }
     ]
+  },
+
+  // TODO/ib - workaround for @loaders.gl/core fs dependency, remove when loaders.gl fixes this
+  node: {
+    fs: 'empty'
   }
 };
