--- test/modules/builder/glb-loader/index.js
@@ -1,5 +1,0 @@
-import './glb-buffer.spec';
-import './pack-json-arrays.spec';
-import './glb-encoder-decoder.spec';
-
-import './glb-encode-parse.spec';
