+++ test/modules/builder/index.js
@@ -1,4 +1,3 @@
-import './glb-loader';
 import './xviz-loader';
 import './builders';
 import './utils';
